Fixing test issues
