+++ cli/test/test_runner_host.py
@@ -1,21 +1,24 @@
 import os
 import unittest
 
+import utils as testutils
 
 from popper.parser import YMLWorkflow
 from popper.runner import WorkflowRunner
-from popper.cli import log as log
 from popper_test import PopperTest
 
+from popper.cli import log as log
 
-class TestHostRunner(unittest.TestCase, PopperTest):
+
+class TestHostHostRunner(unittest.TestCase, PopperTest):
     def setUp(self):
         log.setLevel('CRITICAL')
 
     def test_run(self):
         repo = self.mk_repo()
 
-        with WorkflowRunner(workspace_dir=repo.working_dir) as r:
+        with WorkflowRunner(
+                'docker', 'host', workspace_dir=repo.working_dir) as r:
             wf = YMLWorkflow("""
             version: '1'
             steps:
@@ -48,5 +51,51 @@ class TestHostRunner(unittest.TestCase, PopperTest):
             """)
             wf.parse()
             self.assertRaises(SystemExit, r.run, wf)
+
+        repo.close()
+
+
+class TestHostDockerRunner(unittest.TestCase, PopperTest):
+    def setUp(self):
+        log.setLevel('CRITICAL')
+
+    @unittest.skipIf(os.environ['ENGINE'] != 'docker', 'ENGINE != docker')
+    def test_docker_basic_run(self):
+        repo = self.mk_repo()
+
+        with WorkflowRunner(
+                'docker', 'host', workspace_dir=repo.working_dir) as r:
+            wf = YMLWorkflow("""
+            version: '1'
+            steps:
+            - uses: 'popperized/bin/sh@master'
+              runs: [cat]
+              args: README.md
+            """)
+            wf.parse()
+            r.run(wf)
+
+            wf = YMLWorkflow("""
+            version: '1'
+            steps:
+            - uses: 'docker://alpine:3.9'
+              runs: ['sh', '-c', 'echo $FOO > hello.txt ; pwd']
+              env: {
+                  FOO: bar
+              }
+            """)
+            wf.parse()
+            r.run(wf)
+            with open(os.path.join(repo.working_dir, 'hello.txt'), 'r') as f:
+                self.assertEqual(f.read(), 'bar\n')
+
+            wf = YMLWorkflow("""
+            version: '1'
+            steps:
+            - uses: 'docker://alpine:3.9'
+              runs: 'nocommandisnamedlikethis'
+            """)
+            wf.parse()
+            self.assertRaises(Exception, r.run, wf)
 
         repo.close()
